+++ src/components/MissionCard.tsx
